+++ subscription/util/blocking/competing-consumer-subscription/src/test/java/org/occurrent/subscription/blocking/competingconsumers/CompetingConsumerSubscriptionModelChaosTest.java
@@ -107,7 +107,7 @@ class CompetingConsumerSubscriptionModelChaosTest {
         Retry retry = RetryStrategy
                 .fixed(1000)
                 .retryIf(not(DuplicateCloudEventException.class::isInstance))
-                .onError(((throwable, info) -> log.warn("Detected error (retryCount={}). Error={} - {}. Backoff before retry={}",
+                .onError(((info, throwable) -> log.warn("Detected error (retryCount={}). Error={} - {}. Backoff before retry={}",
                         info.getRetryCount(), throwable.getClass().getSimpleName(), throwable.getMessage(), info.getBackoffBeforeNextRetryAttempt().orElse(Duration.ZERO))));
 
         Thread eventPublishingThread = new Thread(() -> {